+++ sdk/provisioning/Azure.Provisioning/api/Azure.Provisioning.netstandard2.0.cs
@@ -825,9 +825,9 @@ namespace Azure.Provisioning.Primitives
         public System.Collections.Generic.IList<Azure.Provisioning.Expressions.BicepStatement> Statements { get { throw null; } }
         protected internal override System.Collections.Generic.IEnumerable<Azure.Provisioning.Expressions.BicepStatement> Compile() { throw null; }
     }
-    public sealed partial class BicepMetadata
+    public sealed partial class ResourceBicepMetadata
     {
-        public BicepMetadata() { }
+        public ResourceBicepMetadata() { }
         public uint? BatchSize { get { throw null; } set { } }
         public string? Description { get { throw null; } set { } }
         public bool OnlyIfNotExists { get { throw null; } set { } }
@@ -930,7 +930,7 @@ namespace Azure.Provisioning.Primitives
     public abstract partial class ProvisionableResource : Azure.Provisioning.Primitives.NamedProvisionableConstruct
     {
         protected ProvisionableResource(string bicepIdentifier, Azure.Core.ResourceType resourceType, string? resourceVersion = null) : base (default(string)) { }
-        public Azure.Provisioning.Primitives.BicepMetadata BicepMetadata { get { throw null; } }
+        public Azure.Provisioning.Primitives.ResourceBicepMetadata BicepMetadata { get { throw null; } }
         public System.Collections.Generic.IList<Azure.Provisioning.Primitives.ProvisionableResource> DependsOn { get { throw null; } }
         public bool IsExistingResource { get { throw null; } protected set { } }
         public Azure.Core.ResourceType ResourceType { get { throw null; } }